projects/DTOSS-9018-SPIKE-Writing-from-our-laptop-to-ACTUAL-FDP-data-sets/requirements.txt
foundry-platform-sdk==1.9.0
requests==2.32.3
python-dotenv==1.1.0
